+++ changelog.md
@@ -1,2 +1,2 @@
 ### Simple Magnets 1.1.12
-- Initial release of Simple Magnets for Minecraft 1.21.4
+- Initial release of Simple Magnets for Minecraft 1.21.5